stephanrauh/ngx-extended-pdf-viewer#1941 stop trying to write a file in pdf.js if the pdf.js folder isn't there

+++ extract-versions.js
@@ -1,8 +1,10 @@
 const fs = require('fs');
 
-const extendedPdfViewerJsonFile = fs.readFileSync('./projects/ngx-extended-pdf-viewer/package.json');
-const extendedPdfViewerJson = JSON.parse(extendedPdfViewerJsonFile);
-const extendedPdfViewer = extendedPdfViewerJson['version'];
+if (fs.existsSync('../mypdf.js/web/ngx-extended-pdf-viewer-version.js')) {
+  const extendedPdfViewerJsonFile = fs.readFileSync('./projects/ngx-extended-pdf-viewer/package.json');
+  const extendedPdfViewerJson = JSON.parse(extendedPdfViewerJsonFile);
+  const extendedPdfViewer = extendedPdfViewerJson['version'];
 
-const ngxVersion = `export const ngxExtendedPdfViewerVersion = '${extendedPdfViewer}';`;
-fs.writeFileSync('../mypdf.js/web/ngx-extended-pdf-viewer-version.js', ngxVersion);
+  const ngxVersion = `export const ngxExtendedPdfViewerVersion = '${extendedPdfViewer}';`;
+  fs.writeFileSync('../mypdf.js/web/ngx-extended-pdf-viewer-version.js', ngxVersion);
+}
